:recycle: refacto list of events components
