--- client/src/components/2_organisms/newsList/index.ts
@@ -1,1 +1,0 @@
-export * from './newsList'
